requirements.txt
# Core dependencies

groq>=0.4.2
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
chardet>=5.0.0